@@ -4,31 +4,34 @@
   "scripts": {
     "dev": "vite",
     "build": "tsc && vite build",
-    "serve": "vite preview"
+    "serve": "vite preview",
+    "prettier": "prettier --write 'src/**/*.{ts,tsx,js,vue}'"
   },
   "dependencies": {
-    "@reduxjs/toolkit": "^1.7.1",
-    "clsx": "^1.1.1",
-    "evergreen-ui": "^6.6.3",
+    "@emotion/react": "^11.10.4",
+    "@mantine/core": "^5.2.6",
+    "@mantine/hooks": "^5.2.6",
+    "@mantine/notifications": "^5.2.6",
+    "@reduxjs/toolkit": "^1.8.5",
+    "@tanstack/react-table": "^8.5.13",
+    "@tanstack/react-virtual": "^3.0.0-alpha.0",
+    "framer-motion": "^7.2.1",
     "lodash": "^4.17.21",
-    "phosphor-react": "^1.4.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-is": "^17.0.2",
-    "react-redux": "^7.2.6",
-    "use-local-storage-state": "^13.0.0"
+    "npm": "^8.19.1",
+    "phosphor-react": "^1.4.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-redux": "^8.0.2",
+    "react-router-dom": "^6.3.0"
   },
   "devDependencies": {
-    "@types/lodash": "^4.14.178",
-    "@types/node": "^17.0.5",
-    "@types/react": "^17.0.38",
-    "@types/react-dom": "^17.0.11",
-    "@vitejs/plugin-react-refresh": "^1.3.6",
-    "autoprefixer": "^10.4.0",
-    "postcss": "^8.4.5",
-    "prettier": "2.5.1",
-    "tailwindcss": "^3.0.7",
-    "typescript": "^4.5.4",
-    "vite": "^2.7.7"
+    "@types/lodash": "^4.14.184",
+    "@types/node": "^18.7.15",
+    "@types/react": "^18.0.18",
+    "@types/react-dom": "^18.0.6",
+    "@vitejs/plugin-react": "^2.1.0",
+    "prettier": "2.7.1",
+    "typescript": "^4.8.2",
+    "vite": "^3.1.0"
   }
 }
