@@ -1,41 +1,39 @@
 package main
 
 import (
-	"fmt"
 	"os"
 	"path"
 	"path/filepath"
-	"time"
 
 	"github.com/evan-buss/openbooks/server"
+	"github.com/evan-buss/openbooks/util"
 
 	"github.com/spf13/cobra"
 )
 
+var openBrowser = false
 var serverConfig server.Config
 
 func init() {
-	rootCmd.AddCommand(serverCmd)
+	desktopCmd.AddCommand(serverCmd)
 
-	serverConfig.Version = fmt.Sprintf("OpenBooks Server %s", version)
-	serverCmd.Flags().BoolVarP(&serverConfig.Log, "log", "l", false, "Save raw IRC logs for each client connection.")
-	serverCmd.Flags().BoolVarP(&serverConfig.OpenBrowser, "browser", "b", false, "Open the browser on server start.")
-	serverCmd.Flags().StringVarP(&serverConfig.UserName, "name", "n", generateUserName(), "Use a name that isn't randomly generated. One word only.")
 	serverCmd.Flags().StringVarP(&serverConfig.Port, "port", "p", "5228", "Set the local network port for browser mode.")
-	serverCmd.Flags().StringVarP(&serverConfig.DownloadDir, "dir", "d", filepath.Join(os.TempDir(), "openbooks"), "The directory where eBooks are saved when persist enabled.")
-	serverCmd.Flags().BoolVar(&serverConfig.Persist, "persist", false, "Persist eBooks in 'dir'. Default is to delete after sending.")
-	serverCmd.Flags().StringVar(&serverConfig.Basepath, "basepath", "/", `Base path where the application is accessible. For example "/openbooks/".`)
-	serverCmd.Flags().StringVarP(&serverConfig.Server, "server", "s", "irc.irchighway.net", "IRC server to connect to.")
 	serverCmd.Flags().IntP("rate-limit", "r", 10, "The number of seconds to wait between searches to reduce strain on IRC search servers. Minimum is 10 seconds.")
+	serverCmd.Flags().BoolVar(&serverConfig.DisableBrowserDownloads, "no-browser-downloads", false, "The browser won't recieve and download eBook files, but they are still saved to the defined 'dir' path.")
+	serverCmd.Flags().StringVar(&serverConfig.Basepath, "basepath", "/", `Base path where the application is accessible. For example "/openbooks/".`)
+	serverCmd.Flags().BoolVarP(&openBrowser, "browser", "b", false, "Open the browser on server start.")
+	serverCmd.Flags().BoolVar(&serverConfig.Persist, "persist", false, "Persist eBooks in 'dir'. Default is to delete after sending.")
+	serverCmd.Flags().StringVarP(&serverConfig.DownloadDir, "dir", "d", filepath.Join(os.TempDir(), "openbooks"), "The directory where eBooks are saved when persist enabled.")
 }
 
 var serverCmd = &cobra.Command{
 	Use:   "server",
 	Short: "Run OpenBooks in server mode.",
 	Long:  "Run OpenBooks in server mode. This allows you to use a web interface to search and download eBooks.",
-	Run: func(cmd *cobra.Command, args []string) {
+	PreRun: func(cmd *cobra.Command, args []string) {
+		bindGlobalServerFlags(&serverConfig)
 		rateLimit, _ := cmd.Flags().GetInt("rate-limit")
-
+		ensureValidRate(rateLimit, &serverConfig)
 		// If cli flag isn't set (default value) check for the presence of an
 		// environment variable and use it if found.
 		if serverConfig.Basepath == cmd.Flag("basepath").DefValue {
@@ -43,23 +41,14 @@ var serverCmd = &cobra.Command{
 				serverConfig.Basepath = envPath
 			}
 		}
-
-		// If user enters a limit that's too low, set to default of 10 seconds.
-		if rateLimit < 10 {
-			rateLimit = 10
-		}
-
-		serverConfig.SearchTimeout = time.Duration(rateLimit) * time.Second
 		serverConfig.Basepath = sanitizePath(serverConfig.Basepath)
+	},
+	Run: func(cmd *cobra.Command, args []string) {
+		if openBrowser {
+			browserUrl := "http://127.0.0.1:" + path.Join(serverConfig.Port+serverConfig.Basepath)
+			util.OpenBrowser(browserUrl)
+		}
 
 		server.Start(serverConfig)
 	},
-}
-
-func sanitizePath(basepath string) string {
-	cleaned := path.Clean(basepath)
-	if cleaned == "/" {
-		return cleaned
-	}
-	return cleaned + "/"
 }
