@@ -1,12 +1,19 @@
-import { createAction, createSlice, PayloadAction } from "@reduxjs/toolkit";
-import { MessageType, SearchResponse } from "./messages";
+import {
+  createAction,
+  createAsyncThunk,
+  createSlice,
+  PayloadAction
+} from "@reduxjs/toolkit";
 import { addHistoryItem, HistoryItem, updateHistoryItem } from "./historySlice";
-import { AppThunk } from "./store";
+import { MessageType, SearchResponse } from "./messages";
+import { AppDispatch, RootState } from "./store";
 
 interface AppState {
   isConnected: boolean;
+  isSidebarOpen: boolean;
   activeItem: HistoryItem | null;
   username?: string;
+  inFlightDownloads: string[];
 }
 
 const loadActive = (): HistoryItem | null => {
@@ -19,8 +26,10 @@ const loadActive = (): HistoryItem | null => {
 
 const initialState: AppState = {
   isConnected: false,
+  isSidebarOpen: true,
   activeItem: loadActive(),
-  username: undefined
+  username: undefined,
+  inFlightDownloads: []
 };
 
 const stateSlice = createSlice({
@@ -35,6 +44,15 @@ const stateSlice = createSlice({
     },
     setUsername(state, action: PayloadAction<string>) {
       state.username = action.payload;
+    },
+    addInFlightDownload(state, action: PayloadAction<string>) {
+      state.inFlightDownloads.push(action.payload);
+    },
+    removeInFlightDownload(state) {
+      state.inFlightDownloads.shift();
+    },
+    toggleSidebar(state) {
+      state.isSidebarOpen = !state.isSidebarOpen;
     }
   }
 });
@@ -44,10 +62,23 @@ const sendMessage = createAction("socket/send_message", (message: any) => ({
   payload: { message: JSON.stringify(message) }
 }));
 
+const sendDownload = createAsyncThunk(
+  "state/send_download",
+  (book: string, { dispatch }) => {
+    dispatch(addInFlightDownload(book));
+    dispatch(
+      sendMessage({
+        type: MessageType.DOWNLOAD,
+        payload: { book }
+      })
+    );
+  }
+);
+
 // Send a search to the server. Add to query history and set loading.
-const sendSearch =
-  (queryString: string): AppThunk =>
-  (dispatch) => {
+const sendSearch = createAsyncThunk(
+  "state/send_sendSearch",
+  (queryString: string, { dispatch }) => {
     // Send the books search query to the server
     dispatch(
       sendMessage({
@@ -63,12 +94,17 @@ const sendSearch =
     // Add query to item history.
     dispatch(addHistoryItem({ query: queryString, timestamp }));
     dispatch(setActiveItem({ query: queryString, timestamp: timestamp }));
-  };
+  }
+);
 
-const setSearchResults =
-  ({ books, errors }: SearchResponse): AppThunk =>
-  (dispatch, getStore) => {
-    const activeItem = getStore().state.activeItem;
+const setSearchResults = createAsyncThunk<
+  Promise<void>,
+  SearchResponse,
+  { dispatch: AppDispatch; state: RootState }
+>(
+  "state/set_search_results",
+  async ({ books, errors }: SearchResponse, { dispatch, getState }) => {
+    const activeItem = getState().state.activeItem;
     if (activeItem === null) {
       return;
     }
@@ -81,17 +117,18 @@ const setSearchResults =
 
     dispatch(setActiveItem(updatedItem));
     dispatch(updateHistoryItem(updatedItem));
-  };
-const { setActiveItem, setConnectionState, setUsername } = stateSlice.actions;
+  }
+);
 
-export {
-  stateSlice,
-  sendMessage,
+export const {
   setActiveItem,
   setConnectionState,
-  sendSearch,
-  setSearchResults,
-  setUsername
-};
+  setUsername,
+  addInFlightDownload,
+  removeInFlightDownload,
+  toggleSidebar
+} = stateSlice.actions;
+
+export { stateSlice, sendMessage, sendDownload, sendSearch, setSearchResults };
 
 export default stateSlice.reducer;
