@@ -49,7 +49,13 @@ func (server *server) routeMessage(message Request, c *Client) {
 
 // handle ConnectionRequests and either connect to the server or do nothing
 func (c *Client) startIrcConnection(server *server) {
-	core.Join(c.irc, server.config.Server)
+	err := core.Join(c.irc, server.config.Server, server.config.EnableTLS)
+	if err != nil {
+		c.log.Println(err)
+		c.send <- newErrorResponse("Unable to connect to IRC server.")
+		return
+	}
+
 	handler := server.NewIrcEventHandler(c)
 
 	if server.config.Log {
@@ -87,7 +93,7 @@ func (c *Client) sendSearchRequest(s *SearchRequest, server *server) {
 		return
 	}
 
-	core.SearchBook(c.irc, s.Query)
+	core.SearchBook(c.irc, server.config.SearchBot, s.Query)
 	server.lastSearch = time.Now()
 
 	c.send <- newStatusResponse(NOTIFY, "Search request sent.")
