@@ -2,6 +2,7 @@ package util
 
 import (
 	"errors"
+	"fmt"
 	"io"
 	"log"
 	"os"
@@ -15,9 +16,21 @@ var (
 )
 
 func ExtractArchive(archivePath string) (string, error) {
+	// Our path will have a .temp appended to it so we can't rely on the automatic file-extension based archive extractor selection.
+	// This code was taken from the archiver.Walk(archive string, walkFn WalkFunc) error function.
+	// We just remove .temp before trying to find a matching archive extractor.
+	wIface, err := archiver.ByExtension(archivePath[:len(archivePath)-len(".temp")])
+	if err != nil {
+		return "", err
+	}
+	w, ok := wIface.(archiver.Walker)
+	if !ok {
+		return "", fmt.Errorf("format specified by archive filename is not a walker format: %s (%T)", archivePath, wIface)
+	}
+
 	var newPath string
-	err := archiver.Walk(archivePath, func(f archiver.File) error {
-		newPath = filepath.Join(filepath.Dir(archivePath), f.Name())
+	err = w.Walk(archivePath, func(f archiver.File) error {
+		newPath = filepath.Join(filepath.Dir(archivePath), f.Name()+".temp")
 
 		out, err := os.Create(newPath)
 		if err != nil {
@@ -59,6 +72,10 @@ func ExtractArchive(archivePath string) (string, error) {
 // IsArchive returns true if the file at the given path is an archive that can
 // be extracted. Returns false otherwise.
 func IsArchive(path string) bool {
+	if filepath.Ext(path) == ".temp" {
+		path = path[:len(path)-len(".temp")]
+	}
+
 	_, err := archiver.ByExtension(path)
 	return err == nil
 }
